requirements.txt
flask
openai
openrouter
pillow